backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
asyncpg
celery
redis
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
reportlab
jira
PyGithub
rank-bm25
